adjustments to infrastructure

+++ src/infrastructure/two-factor-confirmations/use-cases/create.use-case.ts
@@ -16,20 +16,15 @@ export async function createTwoFactorConfirmationUseCase(
   data: { userId: string }
 ) {
   try {
-    const existingConfirmation = await context.getTwoFactorConfirmationByUser(
-      data.userId
-    );
+    const existingConfirmation = await context.getTwoFactorConfirmationByUser(data.userId);
 
-    if (existingConfirmation)
-      await context.deleteTwoFactorConfirmation(existingConfirmation.id);
+    if (existingConfirmation) await context.deleteTwoFactorConfirmation(existingConfirmation.id);
 
     const confirmation = new TwoFactorConfirmationEntity({
       userId: data.userId,
     });
 
-    await context.createTwoFactorConfirmation(
-      TwoFactorConfirmationToCreateDto(confirmation)
-    );
+    await context.createTwoFactorConfirmation(TwoFactorConfirmationToCreateDto(confirmation));
   } catch (error) {
     console.log("[CREATE_TWO_FACTOR_CONFIRMATION_USE_CASE]: ERROR", error);
   }
